// lib/types/articleType.ts
export interface ArticleList {
  articles: ArticleType[];
}

export interface Article {
  article: ArticleType;
}

export type ArticleType = {
  tagList: string[];
  createdAt: number;
  author: Author;
  description: string;
  title: string;
  body: string;
  slug: string;
  updatedAt: number;
  favoritesCount: number;
  favorited: boolean;
  link: string;
  infographicCode: string
};

export type Author = {
  username: string;
  bio: string;
  image: string;
  following: boolean;
};